+++ docs/make.jl
@@ -13,7 +13,7 @@ makedocs(
 deploydocs(
     repo   = "github.com/diegozea/PairwiseListMatrices.jl.git",
     target = "build",
-    julia  = "0.5",
+    julia  = "0.6",
     deps   = nothing,
     make   = nothing
 )